[fs] fs-cache: When submitting an op, cancel it if the target object is dying
Message-id: <20150407222633.21737.79848.stgit@warthog.procyon.org.uk>
Patchwork-id: 109492
O-Subject: [RHEL7 PATCH 40/50] FS-Cache: When submitting an op, cancel it if the target object is dying [ver #2]
Bugzilla: 1130457 1204964
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

When submitting an operation, prefer to cancel the operation immediately
rather than queuing it for later processing if the object is marked as dying
(ie. the object state machine has reached the KILL_OBJECT state).

Whilst we're at it, change the series of related test_bit() calls into a
READ_ONCE() and bitwise-AND operators to reduce the number of load
instructions (test_bit() has a volatile address).

[Note: For RHEL-7 calls to READ_ONCE() have been converted to ACCESS_ONCE()]

Signed-off-by: David Howells <dhowells@redhat.com>
Reviewed-by: Steve Dickson <steved@redhat.com>
Acked-by: Jeff Layton <jeff.layton@primarydata.com>
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>