[perf] rapl: Fix sysfs_show() initialization for RAPL PMU
Message-id: <1428672113-29331-13-git-send-email-jolsa@redhat.com>
Patchwork-id: 109658
O-Subject: [PATCH RHEL7.2 BZ1209607 012/220] perf/rapl: Fix sysfs_show() initialization for RAPL PMU
Bugzilla: 1209607
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1209607
https://bugzilla.redhat.com/show_bug.cgi?id=1209607

upstream
========
commit 433678bdc6ed39f053c55da96b51de5bf0aeebb1
Author: Stephane Eranian <eranian@google.com>
Date: Tue Jan 13 23:59:53 2015 +0100

description
===========
This patch fixes a problem with the initialization of the
sysfs_show() routine for the RAPL PMU.

The current code was wrongly relying on the EVENT_ATTR_STR()
macro which uses the events_sysfs_show() function in the x86
PMU code. That function itself was relying on the x86_pmu data
structure. Yet RAPL and the core PMU (x86_pmu) have nothing to
do with each other. They should therefore not interact with
each other.

The x86_pmu structure is initialized at boot time based on
the host CPU model. When the host CPU is not supported, the
x86_pmu remains uninitialized and some of the callbacks it
contains are NULL.

The false dependency with x86_pmu could potentially cause crashes
in case the x86_pmu is not initialized while the RAPL PMU is. This
may, for instance, be the case in virtualized environments.

This patch fixes the problem by using a private sysfs_show()
routine for exporting the RAPL PMU events.

Signed-off-by: Rafael Aquini <aquini@redhat.com>