[perf] tools: Add support for 32-bit compatibility VDSOs
Message-id: <1428672113-29331-43-git-send-email-jolsa@redhat.com>
Patchwork-id: 109687
O-Subject: [PATCH RHEL7.2 BZ1209607 042/220] perf tools: Add support for 32-bit compatibility VDSOs
Bugzilla: 1209607
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1209607
https://bugzilla.redhat.com/show_bug.cgi?id=1209607

upstream
========
commit f6832e1720f5cc283703cfe9ccbfb46a3fb6f548
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Thu Oct 23 13:45:23 2014 +0300

description
===========
'perf record' post-processes the event stream  to create a list of
build-ids for object files for which sample events have been recorded.
That results in those object files being recorded in the build-id cache.

In the case of VDSO, perf tools reads it from memory and copies it into
a temporary file, which as decribed above, gets added to the build-id
cache.

Then when the perf.data file is processed by other tools, the build-id
of VDSO is listed in the perf.data file and the VDSO can be read from
the build-id cache.  In that case the name of the map, the short name of
the DSO, and the entry in the build-id cache are all "[vdso]".

However, in the 64-bit case, there also can be 32-bit compatibility
VDSOs.

A previous patch added programs "perf-read-vdso32" and "perf
read-vdsox32".

This patch uses those programs to read the correct VDSO for a thread and
create a temporary file just as for the 64-bit VDSO.

The map name and the entry in the build-id cache are still "[vdso]" but
the DSO short name becomes "[vdso32]" and "[vdsox32]" respectively.

Signed-off-by: Rafael Aquini <aquini@redhat.com>