[perf] tools: Do not attempt to run perf-read-vdso32 if it wasn't built
Message-id: <1428672113-29331-44-git-send-email-jolsa@redhat.com>
Patchwork-id: 109693
O-Subject: [PATCH RHEL7.2 BZ1209607 043/220] perf tools: Do not attempt to run perf-read-vdso32 if it wasn't built
Bugzilla: 1209607
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1209607
https://bugzilla.redhat.com/show_bug.cgi?id=1209607

upstream
========
commit 46b1fa85ff5a2e03423770b3931b97266e8ac6cf
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Thu Oct 23 13:45:24 2014 +0300

description
===========
popen() causes an error message to print if perf-read-vdso32 does not
run.  Avoid that by not trying to run it if it was not built.  Ditto
perf-read-vdsox32.

Signed-off-by: Rafael Aquini <aquini@redhat.com>