[perf] evlist: Do not poll events that use the system_wide flag
Message-id: <1428672113-29331-87-git-send-email-jolsa@redhat.com>
Patchwork-id: 109729
O-Subject: [PATCH RHEL7.2 BZ1209607 086/220] perf evlist: Do not poll events that use the system_wide flag
Bugzilla: 1209607
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1209607
https://bugzilla.redhat.com/show_bug.cgi?id=1209607

upstream
========
commit f90d194a867a5a1db51789b8c4e99bcd196be6bb
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Tue Nov 11 16:16:39 2014 +0200

description
===========
The system_wide flag causes a selected event to be opened always without
a pid.

Consequently it will never get a POLLHUP, but it is used for tracking in
combination with other events, so it should not need to be polled
anyway.

Therefore don't add it for polling.

Signed-off-by: Rafael Aquini <aquini@redhat.com>