[perf] tools: Only override the default :tid comm entry
Message-id: <1428672113-29331-89-git-send-email-jolsa@redhat.com>
Patchwork-id: 109735
O-Subject: [PATCH RHEL7.2 BZ1209607 088/220] perf tools: Only override the default :tid comm entry
Bugzilla: 1209607
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1209607
https://bugzilla.redhat.com/show_bug.cgi?id=1209607

upstream
========
commit a84808083688d82d7f1e5786ccf5df0ff7d448cb
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Tue Nov 11 16:16:41 2014 +0200

description
===========
Events may still be ordered even if there are no timestamps e.g. if the
data is recorded per-thread.

Also synthesized COMM events have a timestamp of zero.

Consequently it is better to keep comm entries even if they have a
timestamp of zero.

However, when a struct thread is created the command string is not known
and a comm entry with a string of the form ":<tid>" is used.

In that case thread->comm_set is false and the comm entry should be
overridden.

Signed-off-by: Rafael Aquini <aquini@redhat.com>