[perf] symbols: Move bfd_demangle stubbing to its only user
Message-id: <1428672113-29331-93-git-send-email-jolsa@redhat.com>
Patchwork-id: 109744
O-Subject: [PATCH RHEL7.2 BZ1209607 092/220] perf symbols: Move bfd_demangle stubbing to its only user
Bugzilla: 1209607
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1209607
https://bugzilla.redhat.com/show_bug.cgi?id=1209607

upstream
========
commit aaba4e12a99cc56fc8614a3f2a3ec6db4fcde76e
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Nov 24 17:10:52 2014 -0300

description
===========
We need to define bfd_demangle() to either a wrapper for
cplus_demangle() or to a stub when NO_DEMANGLE is defined.

That is at odds with using bfd.h for some other reason, as it defines
bfd_demangle() and then if code that wants to use symbol.h, where the
above stubbing/wrapping is done, and bfd.h for other reasons, we end up
with a build error where bfd_demangle() is found to be redefined.

Avoid that by moving the stubbing/wrapping to symbol-elf.c, that is the
only user of such function. If we ever get to a point where there are
more valid users, we can then introduce a header for that.

Signed-off-by: Rafael Aquini <aquini@redhat.com>