[perf] callchain: Make get_srcline fall back to sym+offset
Message-id: <1428672113-29331-94-git-send-email-jolsa@redhat.com>
Patchwork-id: 109737
O-Subject: [PATCH RHEL7.2 BZ1209607 093/220] perf callchain: Make get_srcline fall back to sym+offset
Bugzilla: 1209607
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1209607
https://bugzilla.redhat.com/show_bug.cgi?id=1209607

upstream
========
commit 85c116a6cb91a5c09b7a6c95ffc6a6cbd32cd237
Author: Andi Kleen <ak@linux.intel.com>
Date: Wed Nov 12 18:05:27 2014 -0800

description
===========
When the source line is not found fall back to sym + offset.  This is
generally much more useful than a raw address.

For this we need to pass in the symbol from the caller.

For some callers it's awkward to compute, so we stay at the old
behaviour.

Signed-off-by: Rafael Aquini <aquini@redhat.com>