[perf] evlist: Fixup brown paper bag on "hint" for --mmap-pages cmdline arg
Message-id: <1428672113-29331-130-git-send-email-jolsa@redhat.com>
Patchwork-id: 109773
O-Subject: [PATCH RHEL7.2 BZ1209607 129/220] perf evlist: Fixup brown paper bag on "hint" for --mmap-pages cmdline arg
Bugzilla: 1209607
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1209607
https://bugzilla.redhat.com/show_bug.cgi?id=1209607

upstream
========
commit 0a2138b28fd56eeb3852133691c881409885ee3f
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Fri Dec 12 16:07:47 2014 -0300

description
===========
When failing due to asking for a number of mmap pages that is more than
the max, it was suggesting that an even bigger number of mmap pages
should be specified, doh, au contraire!

Before:

  [acme@ssdandy linux]$ trace -m 128 usleep 1
  Error:	Operation not permitted.
  Hint:	Check /proc/sys/kernel/perf_event_mlock_kb (516 kB) setting.
  Hint:	Tried using 516 kB.
  Hint:	Try using a bigger -m/--mmap-pages value.
  [acme@ssdandy linux]$

After:

  [acme@ssdandy linux]$ trace -m 128 usleep 1
  Error:	Operation not permitted.
  Hint:	Check /proc/sys/kernel/perf_event_mlock_kb (516 kB) setting.
  Hint:	Tried using 516 kB.
  Hint:	Try using a smaller -m/--mmap-pages value.
  [acme@ssdandy linux]$

And to (really) clarify what happens above, when what the user requests
is <= max and even then it fails, a changeset is being made to tell that
this is a per user limit, not per process (in the above example there
was another 'perf trace' running for this user, which was using all the
pages it could use).

Signed-off-by: Rafael Aquini <aquini@redhat.com>