[x86] perf: Fix uncore PCI fixed counter handling
Message-id: <1429539641-10011-9-git-send-email-jolsa@redhat.com>
Patchwork-id: 110922
O-Subject: [PATCH RHEL7.2 BZ1176340 08/42] perf/x86: Fix uncore PCI fixed counter handling
Bugzilla: 1176340
RH-Acked-by: Petr Holasek <pholasek@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1176340
https://bugzilla.redhat.com/show_bug.cgi?id=1176340

upstream
========
commit dbc33f7016747bfce64f3d1e3af63f1251fcbf85
Author: Stephane Eranian <eranian@google.com>
Date: Mon Sep 9 12:53:50 2013 -0700

description
===========
There was a bug in the handling of SNB-EP/IVB-EP uncore PCI
fixed counters, e.g., IMC.

It would cause erratic values to be returned for the IMC
clockticks event. This was due to a bogus hwc->config value
which was then written to PCI config space.

The erratic values can be seen via:

  $ perf stat -a -C 0 -e uncore_imc_0/clockticks/ -I 1000 sleep 10

The fixed counter has most fields marked as reserved with
hw reset values of 0. Yet the kernel was defaulting to a
hwc->config = ~0 and that was causing the issues.

This patch sets the hwc->config values for fixed uncore event
to 0. Now, the values of IMC clockticks is correct.

Signed-off-by: Rafael Aquini <aquini@redhat.com>