[misc] genwqe: Support blocking when DDCB queue is busy
Message-id: <20150421114429.30172.71375.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 111021
O-Subject: [PATCH RHEL7.2 BZ1182052 08/11] GenWQE: Support blocking when DDCB queue is busy
Bugzilla: 1182052
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1182052

Description:
------------
When the GenWQE hardware queue was busy, the driver returned simply
-EBUSY. This caused polling by applications which increased the load
on the already busy system. This change implements the possiblity to
sleep on a waitqueue instead when the DDCB queue is busy. The
requestor is woken up when there is free space on the queue again.
The old way to get -EBUSY is still available if the device is openend
with O_NONBLOCKING. The default is now blocking behavior.

Signed-off-by: Frank Haverkamp <haver@linux.vnet.ibm.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

Upstream Status:
----------------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=1451f414639465995dfc1f820aa1a64723cbd662

----------------
Steve Best

Signed-off-by: Rafael Aquini <aquini@redhat.com>