[kernel] audit: convert status version to a feature bitmap
Message-id: <8feff64e55c16f56979d590e89895a66a3b486b1.1429673629.git.rgb@redhat.com>
Patchwork-id: 111157
O-Subject: [RHEL7.2 PATCH 4/4] audit: convert status version to a feature bitmap
Bugzilla: 1155589
RH-Acked-by: Paul Moore <pmoore@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

The version field defined in the audit status structure was found to have
limitations in terms of its expressibility of features supported.  This is
distict from the get/set features call to be able to command those features
that are present.

Converting this field from a version number to a feature bitmap will allow
distributions to selectively backport and support certain features and will
allow upstream to be able to deprecate features in the future.  It will allow
userspace clients to first query the kernel for which features are actually
present and supported.  Currently, EINVAL is returned rather than EOPNOTSUP,
which isn't helpful in determining if there was an error in the command, or if
it simply isn't supported yet.  Past features are not represented by this
bitmap, but their use may be converted to EOPNOTSUP if needed in the future.

Since "version" is too generic to convert with a #define, use a union in the
struct status, introducing the member "feature_bitmap" unionized with
"version".

Convert existing AUDIT_VERSION_* macros over to AUDIT_FEATURE_BITMAP*
counterparts, leaving the former for backwards compatibility.

Signed-off-by: Richard Guy Briggs <rgb@redhat.com>
[PM: minor whitespace tweaks]
Signed-off-by: Paul Moore <pmoore@redhat.com>
(cherry picked from commit 0288d7183c41c0192d2963d44590f346f4aee917)
[RGB: remove backlog_wait_time, AUDIT_{FEATURE_BITMAP,VERSION}_BACKLOG_WAIT_TIME not backported]
Signed-off-by: Richard Guy Briggs <rgb@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1155589
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=9018109
Test: Tested by me on libvirt kvm rhel7
Signed-off-by: Rafael Aquini <aquini@redhat.com>