[powerpc] powernv: Switch powernv drivers to use machine_xxx_initcall()
Message-id: <20150423145448.9528.53458.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 111200
O-Subject: [PATCH RHEL7.2 BZ1213945 02/13] powerpc/powernv: Switch powernv drivers to use machine_xxx_initcall()
Bugzilla: 1211945
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1211945

Description:
------------
A lot of the code in platforms/powernv is using non-machine initcalls.
That means if a kernel built with powernv support runs on another
platform, for example pseries, the initcalls will still run.

That is usually OK, because the initcalls will check for something in
the device tree or elsewhere before doing anything, so on other
platforms they will usually just return.

But it's fishy for powernv code to be running on other platforms, so
switch them all to be machine initcalls. If we want any of them to run
on other platforms in future they should move to sysdev.

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

Upstream Status:
----------------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=b14726c51ce338ee3abe753fa40428a119c12597

----------------
Steve Best

Signed-off-by: Rafael Aquini <aquini@redhat.com>