[crypto] llvmlinux: Add macro to remove use of VLAIS in crypto code
Message-id: <1430223383-17812-5-git-send-email-nikolay@redhat.com>
Patchwork-id: 111273
O-Subject: [RHEL7.2 PATCH BZ 1173791 04/37] crypto: LLVMLinux: Add macro to remove use of VLAIS in crypto code
Bugzilla: 1173791
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>

From: Behan Webster <behanw@converseincode.com>

Add a macro which replaces the use of a Variable Length Array In Struct (VLAIS)
with a C99 compliant equivalent. This macro instead allocates the appropriate
amount of memory using an char array.

The new code can be compiled with both gcc and clang.

struct shash_desc contains a flexible array member member ctx declared with
CRYPTO_MINALIGN_ATTR, so sizeof(struct shash_desc) aligns the beginning
of the array declared after struct shash_desc with long long.

No trailing padding is required because it is not a struct type that can
be used in an array.

The CRYPTO_MINALIGN_ATTR is required so that desc is aligned with long long
as would be the case for a struct containing a member with
CRYPTO_MINALIGN_ATTR.

If you want to get to the ctx at the end of the shash_desc as before you can do
so using shash_desc_ctx(shash)

Signed-off-by: Behan Webster <behanw@converseincode.com>
Reviewed-by: Mark Charlebois <charlebm@gmail.com>
Acked-by: Herbert Xu <herbert@gondor.apana.org.au>
Cc: Michal Miroslaw <mirqus@gmail.com>
(cherry picked from commit a0a77af14117e388d43b2385d5ecc800e2d6eee9)
Signed-off-by: Rafael Aquini <aquini@redhat.com>