[powerpc] kvm: Implement H_LOGICAL_CI_{LOAD,STORE} in KVM
Message-id: <1430459266-2551-2-git-send-email-dgibson@redhat.com>
Patchwork-id: 111438
O-Subject: [RHEL7.2 kernel PATCH] kvmppc: Implement H_LOGICAL_CI_{LOAD, STORE} in KVM
Bugzilla: 1184290
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>

On POWER, storage caching is usually configured via the MMU - attributes
such as cache-inhibited are stored in the TLB and the hashed page table.

This makes correctly performing cache inhibited IO accesses awkward when
the MMU is turned off (real mode).  Some CPU models provide special
registers to control the cache attributes of real mode load and stores but
this is not at all consistent.  This is a problem in particular for SLOF,
the firmware used on KVM guests, which runs entirely in real mode, but
which needs to do IO to load the kernel.

To simplify this qemu implements two special hypercalls, H_LOGICAL_CI_LOAD
and H_LOGICAL_CI_STORE which simulate a cache-inhibited load or store to
a logical address (aka guest physical address).  SLOF uses these for IO.

However, because these are implemented within qemu, not the host kernel,
these bypass any IO devices emulated within KVM itself.  The simplest way
to see this problem is to attempt to boot a KVM guest from a virtio-blk
device with iothread / dataplane enabled.  The iothread code relies on an
in kernel implementation of the virtio queue notification, which is not
triggered by the IO hcalls, and so the guest will stall in SLOF unable to
load the guest OS.

This patch addresses this by providing in-kernel implementations of the
2 hypercalls, which correctly scan the KVM IO bus.  Any access to an
address not handled by the KVM IO bus will cause a VM exit, hitting the
qemu implementation as before.

Note that a userspace change is also required, in order to enable these
new hcall implementations with KVM_CAP_PPC_ENABLE_HCALL.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
[agraf: fix compilation]
Signed-off-by: Alexander Graf <agraf@suse.de>

(cherry picked from commit 99342cf8044420eebdf9297ca03a14cb6a7085a1)
Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
	arch/powerpc/include/asm/kvm_book3s.h

Trivial textual conflict fixed for downstream.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1184290

Signed-off-by: David Gibson <dgibson@redhat.com>