[redhat] configs: Build accelerated versions of SHA into the kernel
Message-id: <20150504035624.GA24755@gondor.apana.org.au>
Patchwork-id: 111475
O-Subject: [RHEL7 PATCH] crypto: Build accelerated versions of SHA into the kernel
Bugzilla: 1217639
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

RHEL7 Bugzilla #1217639

The crypto layer uses module aliases to ensure that all versions
of a given algorithm are considered.  This means that if you
want to build an algorithm into the kernel statically, then all
implementations of that algorithm that may be used need to be
built in as well to ensure that the best version is chosen.

We were building the generic SHA implementations statically but
not the accelerated versions of SHA, meaning that the accelerated
SHAs never got used.

This patch makes thes the accelerated SHAs built-in as well.

Signed-off-by: Herbert Xu <herbert.xu@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>