[of] Eliminate of_allnodes list
Message-id: <1431374042-16901-1-git-send-email-gduarte@redhat.com>
Patchwork-id: 111831
O-Subject: [RHEL7.2 PATCH BZ 1210533] of: Eliminate of_allnodes list
Bugzilla: 1210533
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

BZ 1210533
Upstream Status: 5063e25a302e6a83f6590d9a06bd5f6400b17430
Build Info: https://brewweb.devel.redhat.com/taskinfo?taskID=9134976
ppc64le: https://brewweb.devel.redhat.com/taskinfo?taskID=9134985
Tested: I did sanity checks on ppc64 and ppc64le. Waiting for further test
results from IBM.

Backport: This patch is slightly modified from upstream as the allnext
pointer cannot be removed from the struct device_node without breaking
KABI.

commit 5063e25a302e6a83f6590d9a06bd5f6400b17430
Author: Grant Likely <grant.likely@linaro.org>
Date:   Fri Oct 3 16:28:27 2014 +0100

    of: Eliminate of_allnodes list

The device tree structure is composed of two lists; the 'allnodes'
list whicha singlely linked depth first list containing every node in
the tree, and theparent->child structure where each parent node has a
singlely linked list of children. All of the data in the allnodes list
can be easily reproduced with the parent->child lists.

This patch is a partial backport of upstream commit 5063e25a30 in
response to observed corruption in the of_allnodes list when
manipulating the device tree dynamically.

Where as that patch completely removes the allnodes list this patch
leaves the of_allnodes list in place including its creation and
linkage at boot when the device tree is intially unflattened. This is
done to reduce the complexity of the backport in light of how out of
sync the driver/of/* tree is. Hence, of_pdt.h, pdt.c, and fdt.c are
left untouched. The key features of this patch is to introduce the
__of_find_all_nodes function, which given a node returns the logical
next node in the device tree, and the for_each_of_allnodes and
for_each_of_allnodes_from macros for iterating over nodes in the
device tree. The remainder of the OF code has been modified to rely on
these functions to iterate the device tree instead of the of_allnodes
list. As such the of_allnodes list exist now only as an artifact.
Finally, it should be noted that of_allnodes will always point to the
root of the device tree and for simplicity a change of name to of_root
is ommitted.

Signed-off-by: Rafael Aquini <aquini@redhat.com>