[kernel] rcu: Do not trace rcu_is_watching() functions
Message-id: <1431432189-8738-4-git-send-email-jolsa@redhat.com>
Patchwork-id: 111880
O-Subject: [PATCH RHEL7.2 BZ1197062 3/4] rcu: Do not trace rcu_is_watching() functions
Bugzilla: 1197062
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Steven Rostedt <rostedt@goodmis.org>

Bugzilla: 1197062
https://bugzilla.redhat.com/show_bug.cgi?id=1197062

upstream
========
commit 9418fb208059d6f916bdb116653e166c5db9456d
Author: Steven Rostedt <rostedt@goodmis.org>
Date:   Mon Nov 4 20:27:36 2013 -0500

description
===========
As perf uses the rcu_read_lock() primitives for recording into its
ring buffer, perf tracing can not be called when RCU in inactive.
With the perf function tracing, there are functions that can be
traced when RCU is not active, and perf must not have its function
callback called when this is the case.

Luckily, Paul McKenney has created a way to detect when RCU is
active or not with the rcu_is_watching() function. Unfortunately,
this function can also be traced, and if that happens it can cause
a bit of overhead for the perf function calls that do the check.
Recursion protection prevents anything bad from happening, but
there is a bit of added overhead for every function being traced that
must detect that the rcu_is_watching() is also being traced.

As rcu_is_watching() is a helper routine and not part of the
critical logic in RCU, it does not need to be traced in order to
debug RCU itself. Add the "notrace" annotation to all the rcu_is_watching()
calls such that we never trace it.

Signed-off-by: Rafael Aquini <aquini@redhat.com>