[perf] tools: Do not check debugfs MAGIC for tracing files
Message-id: <1431774138-14330-54-git-send-email-jolsa@redhat.com>
Patchwork-id: 112277
O-Subject: [PATCH RHEL7.2 BZ1222189 053/327] perf tools: Do not check debugfs MAGIC for tracing files
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit 5693c92660970851e95f769ff27397f5098a6296
Author: Steven Rostedt (Red Hat) <rostedt@goodmis.org>
Date: Mon Feb 2 14:35:02 2015 -0500

description
===========
It's rather strange to be checking the debugfs MAGIC number for the
tracing directory. A system admin may want to have a custom set of
events to trace and it should be allowed to let the admin make a temp
file (even for tracing virtual boxes, this is useful).

Also with the coming tracefs, the files may not even be under debugfs,
so checking the debugfs MAGIC number is pointless.

Signed-off-by: Rafael Aquini <aquini@redhat.com>