[perf] build: Add arch x86 objects building
Message-id: <1431774138-14330-82-git-send-email-jolsa@redhat.com>
Patchwork-id: 112303
O-Subject: [PATCH RHEL7.2 BZ1222189 081/327] perf build: Add arch x86 objects building
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit 5e8c0fb6a95728b852d56c0a9244425d474670c0
Author: Jiri Olsa <jolsa@kernel.org>
Date: Tue Dec 30 14:03:40 2014 +0100

description
===========
Move the x86 arch objects building under build framework to be included
in the libperf build object.

Adding also arch/$(ARCH)/Build files for the rest of the archs. The
reason for this is that in arch/Build we now do:

  +libperf-y += $(ARCH)/

which would make the build to fail on other architectures, because the
build framework requires 'Build' file in nested directories and this
patch adds it only for x86.

Signed-off-by: Rafael Aquini <aquini@redhat.com>